requirements.txt
django>=5.2.0,<5.3.0
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
drf-yasg>=1.21.0
python-dotenv>=1.1.0
pillow>=11.0.0
google-generativeai>=0.8.5
google-auth>=2.40.3
google-cloud-aiplatform>=1.104.0
vertexai>=1.43.0
python-decouple>=3.8
django-storages>=1.14.6
openai>=1.97.1
pyjwt>=2.10.1
djangorestframework-simplejwt>=5.3.0
mcp[cli]>=1.12.0
google-cloud-storage>=2.19.0,<3.0.0
pymysql>=1.1.0
requests>=2.32.4 
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
celery>=5.3.6,<6.0.0
redis>=5.0.3,<6.0.0